Moved the TravelOffersCacheAgent to admin
For great singularity
