+++ admin/app/tools/Store.scala
@@ -38,6 +38,9 @@ trait Store extends Logging with Dates {
   def putDfpLineItemsReport(everything: String) {
     S3.putPublic(dfpLineItemsKey, everything, defaultJsonEncoding)
   }
+  def putCachedTravelOffersFeed(everything: String) {
+    S3.putPublic(travelOffersS3Key, everything, defaultJsonEncoding)
+  }
 
   val now: String = DateTime.now().toHttpDateTimeString
